refactor: move current routes under /api
